weather_streaming/requirements.txt
# Framework để chạy functions trên Google Cloud
functions-framework==3.*

# Thư viện để gọi API
requests==2.*

# Thư viện Google Cloud
google-cloud-pubsub==2.*
google-cloud-bigquery==3.*
pytz